+++ OC-SORT/configs/config.yaml
@@ -3,13 +3,13 @@ save_tubes: True # generated tubes are saved in a separate annotation file in th
 
 Evaluation:
   evaluate: True # do you want to compare generated tubes to the tubes labeled by ROAD?
-  evaluation_path: /road/road_trainval_v1.0.json # path to the ROAD ground truth annotations
+  evaluation_path: '/road/road_trainval_v1.0.json' # path to the ROAD ground truth annotations
 
 Data: 
   # This is for Detections coming out of the detector
-  annotation_path: /road/val1_detections_inactive_merged_gt_format_score_0.1_new.jsonl
+  annotation_path: '/road/detections/eval/flow_color_model_fpn_x101_64x4d_finetune_val1_detections_inactive_merged_gt_format_0.1_new.json'
   ground_truth: False
-  ground_truth_path: /road/road_trainval_v1.0.json # this must be added if ground_truth: False
+  ground_truth_path: '/road/road_trainval_v1.0.json' # this must be added if ground_truth: False
   match_actions: False
   give_video_names: False 
 
@@ -21,7 +21,7 @@ Data:
   action_matching_thresh: 1e-5 # matching threshold for matching a track with their action
 
 OCSORT: # OCSORT parameters
-  det_thresh: 0.8 # threshold for first round matching, detections for second matching are self.det_thresh > score > 0.1
+  det_thresh: 0.4 # threshold for first round matching, detections for second matching are self.det_thresh > score > 0.1
   max_age: 30 # max number of frames in which an object can be gone until that track is considered dead
   min_hits: 3 # minimum number of times a detection must be hit until a track is formed (~vague, should look into)
   iou_threshold: 0.3 # IOU between detection candidates and tracked detection for second round matching
@@ -42,11 +42,17 @@ OCSORT: # OCSORT parameters
 
 Debug:
   debug_video: False # debug tracks by saving a video of it
-  video_name: '2014-06-26-09-53-12_stereo_centre_02' # what video's tracks should we debug on? Must exist in annotation path, can show list in Data.give_video_names
+  video_name: '2015-02-13-09-16-26_stereo_centre_02' # what video's tracks should we debug on? Must exist in annotation path, can show list in Data.give_video_names
 
 Video_Builder:
   video_path: '/road/rgb-images' # path to all the video rgb images
   save_path: '/project/OC-SORT/output' # where the debug video should be loaded to
+
+  compare_detections: True # this option will create a seperate video stream with detection bboxes beside the tracks 
+  # path to the detection annotations
+  detection_path: '/road/flow_color_model_fpn_x101_64x4d_finetune_val1_detections_inactive_merged_gt_format_0.1_new.json'
+
+  # Misc Cv2 params for bbox layout
   bbox_thickness: 2 # pixels thick 
 
   # Title Location: based on the following xy plane
