--- OC-SORT/data/road_video_utils.py
@@ -1,8 +1,0 @@
-import random
-
-def generate_random_colour():
-    r = random.randint(0, 255)
-    g = random.randint(0, 255)
-    b = random.randint(0, 255)
-
-    return (r, g, b)
